[graph] Rearrange 32-bit color.
Both Windows and X11 interpret 32-bit color using masks rather than
byte order. We ned to align with that.

* graph.h (grColor): Update value type.
* graph/grobjs.c (grFindColor),
graph/grfill.c (gr_fill_hline_32),
graph/gblblit.c (grSetTargetPenBrush),
graph/grblit.c (blit_mono_to_rgb32): Updated.

+++ graph/graph.h
@@ -110,7 +110,7 @@
 
   typedef union grColor_
   {
-    long           value;
+    uint32_t       value;
     unsigned char  chroma[4];
 
   } grColor;
